Add back my prior commit (message below).
This change strips the /var/lib/apt/lists directory from the base image. This information makes zero sense to cache because derivatives must use the "apt-get update && apt-get install" pattern to avoid breakage when the lists update anyhow. When this occurs, it costs ~30Mb per layer storing the files, and ultimately this is not useful to the final image in any way 99% of the time.

This change also serves to simply update the basis, since the size of the "apt-get upgrade" is now ~40 Mb.