Add PATH and CMDs.
The image had the correct path set, but I'm not really sure
where it came from. Maybe a default in docker run? There was already
a test.

The image didn't have a CMD set though, and I can't tell where that
was coming from in the old image. It's not in the Dockerfile. Maybe
a default in docker build?